ci: replace CodeQL with golangci-lint and Go-native quality gates
CodeQL (GitHub default setup) was blocking builds. Disable it (done via
repo settings) and replace it with a faster, Go-native quality workflow.

Adds:
- .golangci.yml (schema v2): staticcheck, govet, errcheck, gosec, revive,
  ineffassign, unused, bodyclose, errorlint, misspell, unconvert, copyloopvar,
  nilerr, whitespace + gofmt/goimports formatters. gosec exclusions for
  protocol-mandated weak crypto (MD5/SHA1), desktop-notifier subprocess exec,
  user-file reads, and opt-in insecure TLS are documented inline.
- .github/workflows/quality.yml: golangci-lint (only-new-issues so a future
  lint bump can't retroactively block open PRs), gofmt/goimports diff check,
  go vet, and go test -race -shuffle=on.

Also applies the linter's safe autofixes across the tree (if/else -> switch,
append-variadic, strings.EqualFold, WriteString(Sprintf) -> Fprintf, embedded
field selectors, gofmt) and reverts two misspell false positives on protocol
identifiers ("Africas Talking", Revolt "colour").

No runtime behaviour change; full race-enabled test suite passes.

+++ internal/notify/xbmc_target.go
@@ -82,7 +82,7 @@ func (x *XBMCTarget) BuildRequest(body, title string, notifyType NotifyType) (Re
 		"params": map[string]any{
 			"title":       title,
 			"message":     body,
-			"displaytime": int(x.duration * 1000),
+			"displaytime": x.duration * 1000,
 		},
 		"id": 1,
 	}
